+++ src/app/(auth)/(component)/LoginModal.tsx
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import LoginModalContent from '@/components/auth/login/LoginModalContent';
+import LoginModalContent from '@/app/(auth)/(component)/components/login/LoginModalContent';
 import WhModal from '@/components/elements/modal/WhModal';
 import { ReactComponent as LogoImage } from '../../../../public/assets/images/withaeng-logo-transparent-h42.svg';
