+++ src/app/(auth)/(component)/components/sign-up/CheckEmailModalContent.tsx
