+++ src/app/(auth)/(component)/components/sign-up/SignUpModalContent.tsx
