+++ src/app/(auth)/(component)/components/sign-up/TermModalContent.tsx
